chore: migrate gradle publishing to plugin that supports new central apis
https://central.sonatype.org/publish/publish-portal-gradle/#disclaimer
