+++ .release-please-manifest.json
@@ -1,9 +1,9 @@
 {
   "FlutterMockzilla": "1.0.0",
   "mockzilla-management-ui": "1.2.0",
-  "mockzilla": "2.1.3",
-  "mockzilla-common": "2.1.3",
-  "mockzilla-management": "2.1.3",
+  "mockzilla": "2.2.3",
+  "mockzilla-common": "2.2.3",
+  "mockzilla-management": "2.2.3",
   "FlutterMockzilla/mockzilla": "1.1.0",
   "FlutterMockzilla/mockzilla_android": "1.1.0",
   "FlutterMockzilla/mockzilla_ios": "1.1.2",
