--- build-logic/src/main/kotlin/com/apadmi/mockzilla/VersionUtils.kt
@@ -1,6 +1,0 @@
-package com.apadmi.mockzilla
-
-import org.gradle.api.Project
-
-fun Project.injectedVersion() = if (project.hasProperty("version")) properties["version"].toString()
-    .takeUnless { it.isBlank() || it == "unspecified" } else null
