+++ build.gradle.kts
@@ -11,6 +11,7 @@ plugins {
     alias(libs.plugins.compose) apply false
     alias(libs.plugins.compose.compiler) apply false
     alias(libs.plugins.buildKonfig) apply false
+    alias(libs.plugins.vanniktechPublish) apply false
 }
 
 buildscript {
